+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - イベント集計データをperformancesコレクションにも保管するように調整
+- update @tokyotower/factory
 
 ### Deprecated